refactor: style, code simplification

+++ htdocs/app/api/deleting.php
@@ -5,17 +5,18 @@ session_start();
 if (isset($_GET["documentID"])) {
     $document = DocumentModel::ctorLoad($_GET["documentID"]);
 
-    if ($document->isAuthorized())
+    if ($document && $document->isAuthorized()) {
         $document->delete();
-
-    header("Location: /views/index.phtml");
+        Util::redirect("/views/index.phtml");
+    }
 }
 elseif (isset($_GET["user"])) {
     $user = UserModel::ctorLoad($_SESSION["userID"]);
-    if (! $user)
-        die;
 
-    $user->delete();
-
-    header("Location: /app/forms/log_out.php");
+    if ($user) {
+        $user->delete();
+        Util::redirect("/app/forms/log_out.php");
+    }
 }
+else
+    Util::redirect($_SERVER["HTTP_REFERER"]);
